+++ src/main/java/de/alpharogroup/crypto/model/CharacterDecorator.java
@@ -22,27 +22,15 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-/**
- *
- */
-package de.alpharogroup.crypto.obfuscation.rule;
+package de.alpharogroup.crypto.model;
 
-import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
+import lombok.experimental.SuperBuilder;
 
 /**
- * The unit test class for the class {@link CharacterObfuscationOperationRule}
+ * The class {@link CryptModel} holds a character prefix and a suffix that can decorate an crypt
+ * object
  */
-public class CharacterObfuscationOperationRuleTest
+@SuperBuilder
+public final class CharacterDecorator extends CryptObjectDecorator<Character>
 {
-
-	/**
-	 * Test method for {@link CharacterObfuscationOperationRule}
-	 */
-	@Test
-	public void testWithBeanTester()
-	{
-		BeanTester beanTester = new BeanTester();
-		beanTester.testBean(CharacterObfuscationOperationRule.class);
-	}
 }