+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 ## Change log
 ----------------------
 
+Version 7.2
+-------------
+
+ADDED:
+
+- new generic class that holds a prefix and a suffix that can decorate an crypt object
+- new decorator classes for character and string object created
+- new FUNDING.yml file added for donations added
+
+CHANGED:
+
+- update of parent version to 5.3
+- update of bouncycastle version to 1.64
+- update of guava version to 28.1-jre
+- extended CryptModel class with a List of decorators 
+
 Version 7.1
 -------------